backend/requirements.txt
# ── Web framework ────────────────────────────────────────────────────────────
fastapi[all]>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
python-multipart>=0.0.9

# ── Configuration ─────────────────────────────────────────────────────────────
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.3.0,<3.0.0
python-dotenv>=1.0.1

# ── Database / ORM / Migrations ───────────────────────────────────────────────
sqlalchemy>=2.0.30,<3.0.0
alembic>=1.13.1,<2.0.0
psycopg2-binary>=2.9.9

# ── Task queue ────────────────────────────────────────────────────────────────
celery[redis]>=5.4.0,<6.0.0
redis>=5.0.4,<6.0.0

# ── Gemini / Google AI (new SDK — replaces google-generativeai) ───────────────
google-genai>=1.0.0

# ── OpenAI (optional provider) ────────────────────────────────────────────────
openai>=1.30.0

# ── HuggingFace ───────────────────────────────────────────────────────────────
huggingface-hub>=0.23.0
sentence-transformers>=3.0.0

# ── Vector store ──────────────────────────────────────────────────────────────
faiss-cpu>=1.8.0

# ── Vision / document processing ─────────────────────────────────────────────
opencv-python-headless>=4.9.0
ffmpeg-python>=0.2.0
pillow>=10.3.0
pypdf>=4.2.0
pdf2image>=1.17.0

# ── Sarvam STT ────────────────────────────────────────────────────────────────
sarvamai>=0.1.0
requests>=2.32.0

# ── Supabase / pgvector ───────────────────────────────────────────────────────
supabase>=2.4.0
pgvector>=0.2.5

# ── AWS SDK (S3, Bedrock, DynamoDB, ElastiCache) ──────────────────────────────
boto3>=1.34.0
botocore>=1.34.0

# ── Observability & structured logging ────────────────────────────────────────
python-json-logger>=2.0.7
sentry-sdk[fastapi]>=2.3.0

# ── Auth ──────────────────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4

# ── PDF export (optional — used by report/model-card generators) ──────────────
# weasyprint>=62.0    # uncomment for full-CSS PDF export
# reportlab>=4.2.0    # fallback PDF engine

# ── Dev ───────────────────────────────────────────────────────────────────────
watchdog>=4.0.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=9.0.0
pytest-asyncio>=1.3.0
httpx>=0.25.0
pytest-cov>=7.0.0
factory-boy>=3.3.0
faker>=33.0.0
